chore(test): move and update content block tests
